wifi: make max connection retries configurable
If when connecting to a wifi network, the authentication phase fails,
connman draws the conclusion that the wifi password was wrong. This is a
correct conclusion when wifi reception is good, but this can also occur
on marginal wifi reception.

This adds the setting FavoriteMaxRetries allowing the number of
connection attempts to be changed from the default of 2.

The effect can be observed by temporarily changing the psk of an access
point after the connection has been established, thus causing it to drop.
Restoring the original psk before the maximum retries have been reached
lets the connection be reestablished.

In a location with intermittent reception, allowing unlimited retries is
useful to avoid connman permanently disabling the network.

Signed-off-by: Mans Rullgard <mans@mansr.com>